+++ src/validate.ts
@@ -4,7 +4,8 @@ import type { ScanResult } from '@/scan';
 import * as path from 'node:path';
 import { checkExternalUrl } from './check-external-url';
 import remarkGfm from 'remark-gfm';
-import { readFileFromPath } from './sample';
+import { type PathToUrl, readFileFromPath } from './sample';
+import { resolveUrl } from './utils/url';
 
 const processor = remark().use(remarkGfm);
 
@@ -18,7 +19,7 @@ export type DetectedError = [
   url: string,
   line: number,
   column: number,
-  reason: ErrorReason,
+  reason: ErrorReason | Error,
 ];
 
 export type ValidateConfig = {
@@ -47,6 +48,20 @@ export type ValidateConfig = {
    * @defaultValue false
    */
   checkExternal?: boolean;
+
+  /**
+   * Generate url for markdown files.
+   *
+   * Required for relative url detection.
+   */
+  pathToUrl?: PathToUrl;
+
+  /**
+   * Allowed hrefs, can be:
+   * - a list of hrefs
+   * - a function that returns `true` for allowed href
+   */
+  whitelist?: string[] | ((url: string) => boolean);
 };
 
 export type FileObject = {
@@ -54,6 +69,11 @@ export type FileObject = {
   content: string;
 
   data?: object;
+
+  /**
+   * URL of page, required for relative url detection
+   */
+  url?: string;
 };
 
 /**
@@ -69,20 +89,22 @@ export async function validateFiles(
   const mdExtensions = ['.md', '.mdx'];
 
   async function run(file: string | FileObject): Promise<ValidateError> {
-    const finalFile =
-      typeof file === 'string' ? await readFileFromPath(file) : file;
+    const resolved =
+      typeof file === 'string'
+        ? await readFileFromPath(file, config.pathToUrl)
+        : file;
 
-    if (!mdExtensions.includes(path.extname(finalFile.path))) {
+    if (!mdExtensions.includes(path.extname(resolved.path))) {
       console.warn(
-        `format unsupported: ${finalFile.path}, supported: ${mdExtensions.join(', ')}`,
+        `format unsupported: ${resolved.path}, supported: ${mdExtensions.join(', ')}`,
       );
 
-      return { file: finalFile.path, detected: [] };
+      return { file: resolved.path, detected: [] };
     }
 
     return {
-      file: finalFile.path,
-      detected: await validateMarkdown(finalFile.content, config),
+      file: resolved.path,
+      detected: await validateMarkdown(resolved.content, config, resolved.url),
     };
   }
 
@@ -94,6 +116,7 @@ export async function validateFiles(
 export async function validateMarkdown(
   content: string,
   config: ValidateConfig,
+  baseUrl?: string,
 ) {
   const tree = processor.parse({ value: content });
   const detected: DetectedError[] = [];
@@ -105,11 +128,15 @@ export async function validateMarkdown(
     const pos = node.position;
 
     tasks.push(
-      detect(node.url, config).then((result) => {
-        if (result) {
-          detected.push([node.url, pos.start.line, pos.start.column, result]);
-        }
-      }),
+      detect(node.url, config, baseUrl)
+        .then((result) => {
+          if (result) {
+            detected.push([node.url, pos.start.line, pos.start.column, result]);
+          }
+        })
+        .catch((err: Error) => {
+          detected.push([node.url, pos.start.line, pos.start.column, err]);
+        }),
     );
   });
 
@@ -121,7 +148,10 @@ export async function validateMarkdown(
 export async function detect(
   href: string,
   config: ValidateConfig,
+  baseUrl?: string,
 ): Promise<ErrorReason | undefined> {
+  if (href.startsWith('mailto:')) return;
+
   if (href.match(/https?:\/\//)) {
     if (config.checkExternal) {
       return await checkExternalUrl(href);
@@ -130,10 +160,24 @@ export async function detect(
     return;
   }
 
+  if (config.whitelist) {
+    if (Array.isArray(config.whitelist) && config.whitelist.includes(href))
+      return;
+    if (typeof config.whitelist === 'function' && config.whitelist(href))
+      return;
+  }
+
   const [pathnameWithQuery, fragment] = href.split('#', 2);
-  const [pathname, query] = pathnameWithQuery.split('?', 2);
+  let [pathname, query] = pathnameWithQuery.split('?', 2);
 
   if (pathname.length === 0) return;
+  if (pathname.startsWith('.')) {
+    if (baseUrl) pathname = `/${resolveUrl(baseUrl, pathname)}`;
+    else
+      throw new Error(
+        `relative url ${pathname} cannot be resolved, you need to provide a 'url' property to file`,
+      );
+  }
 
   let meta = config.scanned.urls.get(pathname);
   if (!meta) {